requirements.txt
pandas==2.0.3
nltk==3.8.1
pyenchant==3.2.2
scikit-learn==1.3.0
tensorflow==2.12.0
openpyxl==3.1.2